+++ NEWS.md
@@ -1,5 +1,9 @@
 # rdev 1.10.2
 
+rdev now supports Quarto Documents (`.qmd`) as analysis notebooks:
+
+* Added reference Quarto Document analysis template to `inst/templates/analysis.qmd` (RStudio currently doesn't support [`.qmd` files as document templates](https://github.com/rstudio/rstudio/issues/11316))
+
 * `build_quarto_site()` now supports use of both `Rmd` and/or `qmd` notebooks in `analysis`
 
 * `spell_check_notebooks()` and `update_wordlist_notebooks()` now check both `Rmd` and `qmd` files in the `analysis` directory